refactor(suite-common): move computeBandwidthFeeLevel to wallet-utils
